Loads of changes, can't be arsed to separate them.
Warning, the app will yield a bad access exception immediately; the location delegate is being cleared after the applaunch method is done. It's not the right place to stick it then. On the other hand, I don't want to stick the CLLocationManager instancing in a view because it's literally an application-wide thing we need and it doesn't belong there either - conceptually. Shows I don't know enough cocoa yet. Welp.